replace fixture to spread it for all integration tests
